# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-fiber-boilerplate API documentation",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/indrayyana/go-fiber-boilerplate/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
...
        "/planes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planes de Suscripción"
                ],
                "summary": "Obtener lista de planes de suscripción",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cantidad por página",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Buscar por nombre o descripción",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planes de Suscripción"
                ],
                "summary": "Crear un nuevo plan de suscripción",
                "parameters": [
                    {
                        "description": "Datos del plan",
                        "name": "plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.CreatePlanSuscripcion"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.PlanSuscripcion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.DetailedErrorResponse"
                        }
                    }
                }
            }
        },
        "/planes/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planes de Suscripción"
                ],
                "summary": "Obtener un plan por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del plan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PlanSuscripcion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planes de Suscripción"
                ],
                "summary": "Actualizar un plan por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del plan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados",
                        "name": "plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.UpdatePlanSuscripcion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PlanSuscripcion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planes de Suscripción"
                ],
                "summary": "Eliminar un plan por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del plan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Plan eliminado exitosamente"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.PlanSuscripcion": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "limite_descargas_mensuales": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "precio": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "utils.CommonErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "utils.DetailedErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 422
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Validation failed"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "validation.CreatePlanSuscripcion": {
            "type": "object",
            "required": [
                "activo",
                "limite_descargas_mensuales",
                "nombre",
                "precio"
            ],
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "descripcion": {
                    "type": "string",
                    "maxLength": 255
                },
                "limite_descargas_mensuales": {
                    "type": "integer",
                    "minimum": 1
                },
                "nombre": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "precio": {
                    "type": "number"
                }
            }
        },
...
        "validation.UpdatePlanSuscripcion": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "descripcion": {
                    "type": "string",
                    "maxLength": 255
                },
                "limite_descargas_mensuales": {
                    "type": "integer",
                    "minimum": 1
                },
                "nombre": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "precio": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}